% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{Create App Server}
\usage{
app_server(input, output, session, .app_data, .dir)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny output}

\item{session}{Shiny session}

\item{.app_data}{List with 'data' and 'metadata' elements}

\item{.dir}{Character. Output directory path}
}
\description{
Defines the server logic for the regression comparison app
}
\keyword{internal}
