% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui.R
\name{app_ui}
\alias{app_ui}
\title{Create App UI}
\usage{
app_ui(.app_data)
}
\arguments{
\item{.app_data}{List with 'data' and 'metadata' elements}
}
\value{
Shiny UI
}
\description{
Generates the complete Shiny UI for the regression comparison app
}
\keyword{internal}
