% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_logic.R
\name{apply_filters}
\alias{apply_filters}
\title{Apply Filters to Dataframe}
\usage{
apply_filters(.df, .filter_values)
}
\arguments{
\item{.df}{Dataframe to filter}

\item{.filter_values}{Named list of filter column:value pairs}
}
\value{
Filtered dataframe
}
\description{
Filters dataframe based on selected filter values
}
\keyword{internal}
