% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{create_column_modal}
\alias{create_column_modal}
\title{Create Column Selection Modal}
\usage{
create_column_modal(.tab_name, .model_cols, .current_selection)
}
\arguments{
\item{.tab_name}{Character. Name of the current tab}

\item{.model_cols}{Character vector of model column names}

\item{.current_selection}{Character vector of currently selected columns}
}
\value{
Shiny modal dialog
}
\description{
Creates a modal dialog for selecting visible columns
}
\keyword{internal}
