% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{create_formatter_tab}
\alias{create_formatter_tab}
\title{Create Table Formatter Tab}
\usage{
create_formatter_tab(.app_data)
}
\arguments{
\item{.app_data}{List with 'data' and 'metadata' elements}
}
\value{
Shiny tabPanel
}
\description{
Creates the table formatter tab for filling Excel templates
}
\keyword{internal}
