% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{create_panel_ui}
\alias{create_panel_ui}
\title{Create Panel UI}
\usage{
create_panel_ui(.tab_name, .metadata)
}
\arguments{
\item{.tab_name}{Character. Name of the current tab}

\item{.metadata}{List. Metadata for the current table}
}
\value{
Shiny UI element
}
\description{
Creates the complete UI for one panel
}
\keyword{internal}
