% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{create_significance_tab}
\alias{create_significance_tab}
\title{Create Significance Overview Tab}
\usage{
create_significance_tab(.app_data)
}
\arguments{
\item{.app_data}{List with 'data' and 'metadata' elements}
}
\value{
Shiny tabPanel
}
\description{
Creates the significance visualization tab
}
\keyword{internal}
