% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{create_table_tab}
\alias{create_table_tab}
\title{Creates a complete tab with one table for viewing regressions}
\usage{
create_table_tab(.tab_name, .metadata)
}
\arguments{
\item{.tab_name}{Character. Name of the tab (from list names)}

\item{.metadata}{List. Metadata for this table}
}
\value{
Shiny tabPanel
}
\description{
Creates a complete tab with one table for viewing regressions
}
\keyword{internal}
