% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detect_column_types}
\alias{detect_column_types}
\title{Detect Column Types in Dataframe}
\usage{
detect_column_types(.df)
}
\arguments{
\item{.df}{Dataframe with 'term' column}
}
\value{
Named list with 'filter_cols', 'term_col', 'model_cols', 'term_idx'
}
\description{
Identifies which columns are filters, term, and models
}
\keyword{internal}
