% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_fill.R, R/template_helpers.R
\name{get_available_columns}
\alias{get_available_columns}
\title{Get Available Columns for Table}
\usage{
get_available_columns(.app_data, .table_name)

get_available_columns(.app_data, .table_name)
}
\arguments{
\item{.app_data}{List with 'data' and 'metadata' elements}

\item{.table_name}{Character. Name of table}
}
\value{
Character vector of column names

Character vector of column names
}
\description{
Returns available model columns for a given table

Returns available model columns for a given table
}
\keyword{internal}
