% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_fill.R, R/template_helpers.R
\name{get_available_rows}
\alias{get_available_rows}
\title{Get Available Rows for Table}
\usage{
get_available_rows(.app_data, .table_name)

get_available_rows(.app_data, .table_name)
}
\arguments{
\item{.app_data}{List with 'data' and 'metadata' elements}

\item{.table_name}{Character. Name of table}
}
\value{
Character vector of row terms

Character vector of row terms
}
\description{
Returns available row terms for a given table

Returns available row terms for a given table
}
\keyword{internal}
