% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_fill.R, R/template_helpers.R
\name{get_available_tables}
\alias{get_available_tables}
\title{Get Available Dashboard Tables}
\usage{
get_available_tables(.app_data)

get_available_tables(.app_data)
}
\arguments{
\item{.app_data}{List with 'data' and 'metadata' elements}
}
\value{
Character vector of table names

Character vector of table names
}
\description{
Returns list of available tables from app data

Returns list of available tables from app data
}
\keyword{internal}
