% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_default_filter_values}
\alias{get_default_filter_values}
\title{Get Default Filter Values}
\usage{
get_default_filter_values(.df, .filter_cols)
}
\arguments{
\item{.df}{Dataframe}

\item{.filter_cols}{Character vector of filter column names}
}
\value{
Named list where names are filter columns and values are first unique value
}
\description{
Gets the first unique value for each filter column (default selection)
}
\keyword{internal}
