% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_logic.R
\name{get_filtered_data}
\alias{get_filtered_data}
\title{Get Filtered Data}
\usage{
get_filtered_data(.df, .filters)
}
\arguments{
\item{.df}{Dataframe to filter}

\item{.filters}{Named list of filter values}
}
\value{
Filtered dataframe
}
\description{
Main filtering function that applies all filters
}
\keyword{internal}
