% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_first_value}
\alias{get_first_value}
\title{Get First Unique Value}
\usage{
get_first_value(.x)
}
\arguments{
\item{.x}{Vector}
}
\value{
First unique value
}
\description{
Returns the first unique value from a vector (used for default filter values)
}
\keyword{internal}
