% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_unique_filter_values}
\alias{get_unique_filter_values}
\title{Get Unique Filter Values}
\usage{
get_unique_filter_values(.df, .filter_cols)
}
\arguments{
\item{.df}{Dataframe}

\item{.filter_cols}{Character vector of filter column names}
}
\value{
Named list where names are filter columns and values are unique values
}
\description{
Extracts unique values for each filter column
}
\keyword{internal}
