% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_storage.R
\name{list_available_mappings}
\alias{list_available_mappings}
\title{List Available Mappings}
\usage{
list_available_mappings(.dir)
}
\arguments{
\item{.dir}{Character. Directory containing mappings}
}
\value{
Character vector of mapping file paths
}
\description{
Lists all saved mapping files in a directory
}
\keyword{internal}
