% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_from_excel}
\alias{load_from_excel}
\title{Load Data from Excel File}
\usage{
load_from_excel(.path)
}
\arguments{
\item{.path}{Character. Path to .xlsx file}
}
\value{
Named list of dataframes (one per sheet)
}
\description{
Loads multiple sheets from an Excel file as a named list of dataframes
}
\keyword{internal}
