% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_from_rds}
\alias{load_from_rds}
\title{Load Data from RDS File}
\usage{
load_from_rds(.path)
}
\arguments{
\item{.path}{Character. Path to .rds file}
}
\value{
Named list of dataframes
}
\description{
Loads a named list of dataframes from an .rds file
}
\keyword{internal}
