% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_storage.R
\name{load_mapping_config}
\alias{load_mapping_config}
\title{Load Mapping Configuration}
\usage{
load_mapping_config(.path)
}
\arguments{
\item{.path}{Character. Path to mapping JSON file}
}
\value{
List with mapping configuration
}
\description{
Loads a mapping configuration from a JSON file
}
\keyword{internal}
