% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_regression_data}
\alias{load_regression_data}
\title{Load Regression Data from File}
\usage{
load_regression_data(.path)
}
\arguments{
\item{.path}{Character. Path to input file (.rds or .xlsx)}
}
\value{
Named list of dataframes
}
\description{
Loads regression data from either .rds or .xlsx file format
}
\keyword{internal}
