% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{prepare_app_data}
\alias{prepare_app_data}
\title{Prepare Data for App}
\usage{
prepare_app_data(.data_list)
}
\arguments{
\item{.data_list}{Named list of dataframes}
}
\value{
List with 'data' and 'metadata' elements
}
\description{
Processes loaded data and extracts metadata for each table
}
\keyword{internal}
