% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_storage.R
\name{save_mapping_config}
\alias{save_mapping_config}
\title{Save Mapping Configuration}
\usage{
save_mapping_config(.mapping, .dir, .name)
}
\arguments{
\item{.mapping}{List with mapping configuration}

\item{.dir}{Character. Directory to save mapping}

\item{.name}{Character. Name for the mapping file}
}
\value{
Character. Path to saved file
}
\description{
Saves the current mapping configuration to a JSON file
}
\keyword{internal}
