% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setup_output_dir}
\alias{setup_output_dir}
\title{Set Up Output Directory}
\usage{
setup_output_dir(.dir = NULL)
}
\arguments{
\item{.dir}{Character or NULL. Output directory path}
}
\value{
Character. Path to the output directory
}
\description{
Creates output directory if it doesn't exist, or uses a temporary directory
}
\keyword{internal}
