% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_display.R
\name{render_comparison_table}
\alias{render_comparison_table}
\title{Render Comparison Table}
\usage{
render_comparison_table(
  .df,
  .term_col,
  .bold_rows = NULL,
  .border_top_rows = NULL,
  .border_bottom_rows = NULL
)
}
\arguments{
\item{.df}{Dataframe to display (should only contain term + model columns)}

\item{.term_col}{Character. Name of term column}

\item{.bold_rows}{Character vector. Row term values to make bold (NULL if none)}

\item{.border_top_rows}{Character vector. Row term values to add top borders (NULL if none)}

\item{.border_bottom_rows}{Character vector. Row term values to add bottom borders (NULL if none)}
}
\value{
reactable object
}
\description{
Creates a reactable table for display in the app
}
\keyword{internal}
