% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_display.R
\name{render_comparison_table}
\alias{render_comparison_table}
\title{Render Comparison Table}
\usage{
render_comparison_table(.df, .term_col)
}
\arguments{
\item{.df}{Dataframe to display (should only contain term + model columns)}

\item{.term_col}{Character. Name of term column}
}
\value{
reactable object
}
\description{
Creates a reactable table for display in the app
}
\keyword{internal}
